add comment for aliasUpdate usage
Use aliasUpdate=true if you use sling:alias or vanity urls. This will send a flush request to dispatcher which contians both the original path and the vanity path with aliases. Otherwise only the original path will be flushed in dispatcher cache and the aliased path will remain in cache